add deep probe
