+++ src/probe/deep.rs
@@ -1,0 +1,123 @@
+use crate::format_context::FormatContext;
+use log::LevelFilter;
+use stainless_ffmpeg_sys::*;
+use std::cmp;
+use std::fmt;
+
+#[derive(Debug, Deserialize, PartialEq, Serialize)]
+pub struct DeepProbe {
+  #[serde(skip_serializing)]
+  filename: String,
+  pub result: Option<DeepProbeResult>,
+}
+
+#[derive(Debug, Deserialize, PartialEq, Serialize)]
+pub struct DeepProbeResult {
+  #[serde(default)]
+  streams: Vec<StreamProbeResult>
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct StreamProbeResult {
+  count_packets: usize,
+  min_packet_size: i32,
+  max_packet_size: i32,
+}
+
+impl fmt::Display for DeepProbeResult {
+  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    for (index, stream) in self.streams.iter().enumerate() {
+      writeln!(f, "\n{:30} : {:?}", "Stream Index", index)?;
+      writeln!(f, "{:30} : {:?}", "Number of packets", stream.count_packets)?;
+      writeln!(f, "{:30} : {:?}", "Minimum packet size", stream.min_packet_size)?;
+      writeln!(f, "{:30} : {:?}", "Maximum packet size", stream.max_packet_size)?;
+    }
+    Ok(())
+  }
+}
+
+impl StreamProbeResult {
+  pub fn new() -> Self {
+    StreamProbeResult {
+      count_packets: 0,
+      min_packet_size: std::i32::MAX,
+      max_packet_size: std::i32::MIN,
+    }
+  }
+}
+
+impl DeepProbe {
+  pub fn new(filename: &str) -> Self {
+    DeepProbe {
+      filename: filename.to_owned(),
+      result: None,
+    }
+  }
+
+  pub fn process(&mut self, log_level: LevelFilter) -> Result<(), String> {
+    let av_log_level =
+      match log_level {
+        LevelFilter::Error => AV_LOG_ERROR,
+        LevelFilter::Warn => AV_LOG_WARNING,
+        LevelFilter::Info => AV_LOG_INFO,
+        LevelFilter::Debug => AV_LOG_DEBUG,
+        LevelFilter::Trace => AV_LOG_TRACE,
+        LevelFilter::Off => AV_LOG_QUIET,
+      };
+
+    unsafe {
+      av_log_set_level(av_log_level);
+    }
+
+    let mut context = FormatContext::new(&self.filename).unwrap();
+    if context.open_input().is_err() {
+      self.result = None;
+      context.close_input();
+      return Ok(());
+    }
+
+    let mut streams = vec![];
+    streams.resize(context.get_nb_streams() as usize, StreamProbeResult::new());
+    loop {
+      match context.next_packet() {
+        Ok(packet) => {
+          unsafe {
+            let stream_index = (*packet.packet).stream_index as usize;
+            let packet_size = (*packet.packet).size;
+
+            streams[stream_index].count_packets += 1;
+            streams[stream_index].min_packet_size = cmp::min(packet_size, streams[stream_index].min_packet_size);
+            streams[stream_index].max_packet_size = cmp::max(packet_size, streams[stream_index].max_packet_size);
+          }
+        },
+        Err(_) => {
+          break;
+        }
+      }
+    }
+
+    self.result = Some(DeepProbeResult {
+      streams
+    });
+
+    context.close_input();
+    Ok(())
+  }
+}
+
+#[test]
+fn deep_probe_mxf_sample() {
+  use serde_json;
+  use std::fs::File;
+  use std::io::prelude::*;
+
+  let mut probe = DeepProbe::new("tests/PAL_1080i_MPEG_XDCAM-HD_colorbar.mxf");
+  probe.process(LevelFilter::Error).unwrap();
+
+  let mut file = File::open("tests/deep_probe.json").unwrap();
+  let mut contents = String::new();
+  file.read_to_string(&mut contents).unwrap();
+
+  let reference : DeepProbe = serde_json::from_str(&contents).unwrap();
+  assert_eq!(probe, reference);
+}
